watchdog: ts4800: add driver for TS-4800 watchdog
This watchdog is instantiated in a FPGA that is memory mapped. It is
made of only one register, called the feed register. Writing to this
register will re-arm the watchdog for a given time (and enable it if it
was disable). It can be disabled by writing a special value into the
feed register.

It is instantiated in a larger block called syscon. On TS-4800, the
watchdog register is at offset 0xE from syscon's base, but other
Technologic System's boards instantiate it at different offsets. This
offset must be provided in a device tree node.

Signed-off-by: Damien Riegel <damien.riegel@savoirfairelinux.com>